Get climate data prepared for VIMC and stored locally

Load climate data prepared for VIMC from a local directory.

Usage

load_local_vimc_climate(
  country,
  date_range,
  data_location,
  data_source = NULL,
  admin_level = c(0, 1, 2, 3)
)

Arguments

country

A string for the country name, or for the country ISO 2- or 3- character code. E.g. one of "Canada", "CA", or "CAN". Input is checked against names and codes in countrycode.

date_range

A two element vector of <Date>s, giving the interval for which to get data.

data_location

A location relative to the current working directory to search for the data. Data for country are expected to live in data_location/XYZ as Parquet files, where XYZ is the ISO 3-character code for country. Defaults to the working directory.

data_source

A string giving the climate data to search for. Names must match exactly.

admin_level

A numeric giving the GADM admin level for which to get data. Only one unit may be passed at a time.

Value

A <data.frame> of climate data.