Locate local data files

Locate local data files

Usage

locate_parquet_file(data_location, country_iso3c, admin_level, data_source)

Arguments

data_location

A location relative to the current working directory to search for the data. Data for country are expected to live in data_location/XYZ as Parquet files, where XYZ is the ISO 3-character code for country. Defaults to the working directory.

country_iso3c

Country ISO 3 character code. Passed from top-level function load_local_vimc_climate() after a call to validate_country().

admin_level

A numeric giving the GADM admin level for which to get data. Only one unit may be passed at a time.

data_source

A string giving the climate data to search for. Names must match exactly.

Value

The location of a local data Parquet file. Has a number of side-effects to inform users when data_location, the county data directory within data_location, or the file corresponding to the data requested are not found.